Add PathFinder.java
You are a software engineer developing a city- and intercity-planning software to used by urban
planning agencies (e.g., road-works department, emergency management departments). As a result, the software provides answers to different type of queries such as shortest distances between
intersections in the city, which allows for planning of bus-routes.
It has been decided that the intersections in the city should be identified using an integer
identifier and their locations are described using coordinates. Furthermore, if there is a roadway
between two intersections, the distance or length of the roadway can be approximated to be the
Euclidean distance between the intersections.